requirements.txt
openai>=1.0.0
numpy>=1.24
faiss-cpu>=1.7.0
python-dotenv>=1.0.0
pypdf>=3.0
